move get world response into handler.
